put all myClass controllers to another class
