+++ src/main/java/com/inva/hipstertest/service/mapper/ParentMapper.java
@@ -4,12 +4,13 @@ import com.inva.hipstertest.domain.*;
 import com.inva.hipstertest.service.dto.ParentDTO;
 
 import org.mapstruct.*;
+
 import java.util.List;
 
 /**
  * Mapper for the entity Parent and its DTO ParentDTO.
  */
-@Mapper(componentModel = "spring", uses = {UserMapper.class, })
+@Mapper(componentModel = "spring", uses = {UserMapper.class,})
 public interface ParentMapper {
 
     @Mapping(source = "user.id", target = "userId")
@@ -24,6 +25,7 @@ public interface ParentMapper {
     Parent parentDTOToParent(ParentDTO parentDTO);
 
     List<Parent> parentDTOsToParents(List<ParentDTO> parentDTOs);
+
     /**
      * generating the fromId for all mappers if the databaseType is sql, as the class has relationship to it might need it, instead of
      * creating a new attribute to know if the entity has any relationship from some other entity
